requirements.txt
numpy
Pillow
sqlalchemy
